+++ movies/src/Featured.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Featured = (props) => {
+  return (
+    <div className="container">
+      <h1>Futures!! {props.match.params.id}</h1>
+    </div>
+  );
+}
+
+export default Featured;
